+++ engine/src/test/java/net/jqwik/engine/properties/arbitraries/DefaultTypeArbitraryTests.java
@@ -7,6 +7,7 @@ import org.assertj.core.api.*;
 import net.jqwik.api.*;
 import net.jqwik.api.arbitraries.*;
 import net.jqwik.api.providers.*;
+import net.jqwik.engine.properties.*;
 
 import static net.jqwik.engine.properties.ArbitraryTestHelper.*;
 
@@ -30,7 +31,7 @@ class DefaultTypeArbitraryTests {
 
 		TypeArbitrary<String> typeArbitrary =
 			new DefaultTypeArbitrary<>(String.class)
-				.use(getClass().getDeclaredMethod("stringFromNoParams"));
+				.use(Samples.class.getDeclaredMethod("stringFromNoParams"));
 
 		assertAllGenerated(
 			typeArbitrary.generator(1000),
@@ -43,7 +44,7 @@ class DefaultTypeArbitraryTests {
 
 		TypeArbitrary<String> typeArbitrary =
 			new DefaultTypeArbitrary<>(String.class)
-				.use(getClass().getDeclaredMethod("stringFromNoParams"))
+				.use(Samples.class.getDeclaredMethod("stringFromNoParams"))
 				.use(String.class.getConstructor());
 
 		RandomGenerator<String> generator = typeArbitrary.generator(1000);
@@ -57,41 +58,98 @@ class DefaultTypeArbitraryTests {
 	}
 
 	@Example
-	void typeArbitraryWithoutUseFailsOnGeneration() throws NoSuchMethodException {
-		TypeArbitrary<String> typeArbitrary = new DefaultTypeArbitrary<>(String.class);
+	@Disabled("currently failing") //TODO: fix
+	void exceptionsDuringCreationAreIgnored() throws NoSuchMethodException {
+		TypeArbitrary<String> typeArbitrary =
+			new DefaultTypeArbitrary<>(String.class)
+				.use(Samples.class.getDeclaredMethod("stringWithRandomException"));
 
-		Assertions.assertThatThrownBy(
-			() -> typeArbitrary.generator(1000)
-		).isInstanceOf(JqwikException.class);
+		RandomGenerator<String> generator = typeArbitrary.generator(1000);
+
+		assertAllGenerated(
+			generator,
+			aString -> {
+				return aString.equals("a string");
+			}
+		);
 	}
-
 
 	@Example
-	void nonStaticMethodsAreNotSupported() {
-		Assertions.assertThatThrownBy(
-			() -> new DefaultTypeArbitrary<>(String.class)
-					  .use(getClass().getDeclaredMethod("nonStaticMethod"))
-		).isInstanceOf(JqwikException.class);
+	void useConstructorWithOneParameter() throws NoSuchMethodException {
+
+		TypeArbitrary<Person> typeArbitrary =
+			new DefaultTypeArbitrary<>(Person.class)
+				.use(Person.class.getConstructor(String.class));
+
+		assertAllGenerated(
+			typeArbitrary.generator(1000),
+			aPerson -> aPerson.toString().length() <= 10
+		);
 	}
 
-	@Example
-	void creatorWithWrongReturnTypeIsNotSupported() {
-		Assertions.assertThatThrownBy(
-			() -> new DefaultTypeArbitrary<>(TypeUsage.of(List.class, TypeUsage.of(int.class)))
-					  .use(getClass().getDeclaredMethod("listOfStringsFromNoParams"))
-		).isInstanceOf(JqwikException.class);
+	@Group
+	class ConfigurationErrors {
+		@Example
+		void typeArbitraryWithoutUseFailsOnGeneration() throws NoSuchMethodException {
+			TypeArbitrary<String> typeArbitrary = new DefaultTypeArbitrary<>(String.class);
+
+			Assertions.assertThatThrownBy(
+				() -> typeArbitrary.generator(1000)
+			).isInstanceOf(JqwikException.class);
+		}
+
+		@Example
+		void nonStaticMethodsAreNotSupported() {
+			Assertions.assertThatThrownBy(
+				() -> new DefaultTypeArbitrary<>(String.class)
+						  .use(Samples.class.getDeclaredMethod("nonStaticMethod"))
+			).isInstanceOf(JqwikException.class);
+		}
+
+		@Example
+		void creatorWithWrongReturnTypeIsNotSupported() {
+			Assertions.assertThatThrownBy(
+				() -> new DefaultTypeArbitrary<>(TypeUsage.of(List.class, TypeUsage.of(int.class)))
+						  .use(Samples.class.getDeclaredMethod("listOfStringsFromNoParams"))
+			).isInstanceOf(JqwikException.class);
+		}
+
 	}
 
-	private static String stringFromNoParams() {
-		return "a string";
+	private static class Samples {
+
+		private static String stringFromNoParams() {
+			return "a string";
+		}
+
+		private static String stringWithRandomException() {
+			if (SourceOfRandomness.current().nextDouble() > 0.5) {
+				throw new AssertionError();
+			}
+			return "a string";
+		}
+
+		private static List<String> listOfStringsFromNoParams() {
+			return Arrays.asList("a", "b");
+		}
+
+		private String nonStaticMethod() {
+			return "a string";
+		}
 	}
 
-	private static List<String> listOfStringsFromNoParams() {
-		return Arrays.asList("a", "b");
-	}
+	private static class Person {
+		private final String name;
 
-	private String nonStaticMethod() {
-		return "a string";
+		public Person(String name) {
+			if (name.length() > 10) name = name.substring(0, 10);
+			this.name = name;
+		}
+
+		@Override
+		public String toString() {
+			return name;
+		}
 	}
 
 }